% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.model.R
\docType{package}
\name{aggregate.model}
\alias{aggregate.model-package}
\alias{aggregate.model}
\title{\code{aggregate.model} package}
\description{
The Package for the Aggregate Model
}
\details{
See the README on
\href{https://cran.r-project.org/package=aggregate.model/README.html}{CRAN}
or \href{https://github.com/moritzpschwarz/aggregate.model#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://moritzschwarz.org/getspanel/}
}

}
\author{
\strong{Maintainer}: Moritz Schwarz \email{moritz.schwarz@scmo.eu} (\href{https://orcid.org/0000-0003-0340-3780}{ORCID})

Authors:
\itemize{
  \item Jonas Kurle (\href{https://orcid.org/0000-0003-2197-2012}{ORCID})
}

}
