% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.arx.isat.osem.R
\name{as.arx.isat.osem}
\alias{as.arx.isat.osem}
\title{OSEM adaptation of as.arx.isat()}
\usage{
as.arx.isat.osem(object, ...)
}
\arguments{
\item{object}{An \code{isat} object}

\item{...}{Further arguments}
}
\value{
An \code{arx} object
}
\description{
OSEM adaptation of as.arx.isat()
}
