% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{build_candidate_design}
\alias{build_candidate_design}
\title{Build a candidate design matrix by removing and adding indicator terms}
\usage{
build_candidate_design(model_object, remove_terms, add_terms)
}
\description{
Build a candidate design matrix by removing and adding indicator terms
}
\keyword{internal}
