% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_identities.R
\name{calculate_identities}
\alias{calculate_identities}
\title{Calculate values for identity/definition variables not available in Eurostat}
\usage{
calculate_identities(specification, data, dictionary = NULL)
}
\arguments{
\item{specification}{A tibble or data.frame as returned by
\code{\link{translate_variables}}.}

\item{data}{A tibble or data.frame containing the downloaded or locally
loaded data as returned by \code{\link{load_or_download_variables}}.}

\item{dictionary}{A tibble or data.frame storing the model variable names in
a column named 'model_varname' and the corresponding Eurostat variable code
in 'eurostat_code'. When \code{NULL}, the \link[=dict]{default dictionary}
is used.}
}
\value{
Returns the original tibble or data.frame \code{data} with added
values for the identity/definition variables that were not available in
Eurostat.
}
\description{
Calculates true values for variables that are given by identity or definition but are not directly available from Eurostat.
}
