% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frequencies.R
\name{check_frequencies}
\alias{check_frequencies}
\title{An internal function that checks the frequencies of the model values}
\usage{
check_frequencies(full_data, quiet)
}
\arguments{
\item{full_data}{The full dataset that is used in the model}

\item{quiet}{Logical with default = FALSE. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}
}
\value{
A list with the updated full_data and the frequency of the data
}
\description{
This function checks and deals with different types of mixed frequency models. When there is a mixed frequency within a variable, it tries to remove
}
