% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{classify_variables}
\alias{classify_variables}
\title{Classify variables}
\usage{
classify_variables(specification)
}
\arguments{
\item{specification}{A specification to be classified.}
}
\description{
Classifies variables according to their state in the model: exogenous (x),
endogenous by modelling (n), and endogenous by definition/identity (d).
}
\details{
NOTE: exported for now, later will be internal
}
