% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Prepare the clean dataframe for modelling}
\usage{
clean_data(
  raw_data,
  max.ar = 4,
  max.dl = 2,
  trend = TRUE,
  opts_df,
  module,
  use_logs
)
}
\arguments{
\item{raw_data}{A tibble or data.frame with the y variable and the x
variables. Needs to have a column called 'time', which is of class
\code{\link[base:Dates]{Date}}. Variable names need to be in column
'na_item', and values in column 'values'.}

\item{max.ar}{Integer. The maximum number of lags to use for the AR terms. as well as for the independent variables.}

\item{max.dl}{Integer. The maximum number of lags to use for the independent variables (the distributed lags).}

\item{trend}{Logical. Should a trend be added? Default is TRUE.}

\item{opts_df}{Internal object containing detailed options and information on individual modules.}

\item{module}{A row of the specification table.}

\item{use_logs}{To decide whether to log any variables. Must be one of
'both', 'y', 'x', or 'none'. Default is 'both'.}
}
\value{
A tibble with the cleaned data.
}
\description{
Prepare the clean dataframe for modelling
}
