% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_forecast_recipe.R
\name{compile_forecast_recipe}
\alias{compile_forecast_recipe}
\title{Compile the information required to forecast an estimated OSEM module}
\usage{
compile_forecast_recipe(
  model_object,
  model_form,
  dep_var_basename,
  x_vars_basename,
  use_logs,
  transformations,
  term_specs,
  lag_only_vars = character(0)
)
}
\description{
Compile the information required to forecast an estimated OSEM module
}
\keyword{internal}
