% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionaries.R
\name{create_dictionaries}
\alias{create_dictionaries}
\title{Internal function to create the dictionaries
This script contains a separate function called generate_dictionary_script. This can be used to read in a modified dictionary and will then give out the code to insert into this function.}
\usage{
create_dictionaries(single_df = FALSE)
}
\arguments{
\item{single_df}{Logical. If TRUE, the function will return a single data frame with all dictionaries. If FALSE, the function will return a list of data frames.}
}
\value{
Either a list of dictionaries or a data.frame with a single dictionary.
}
\description{
Internal function to create the dictionaries
This script contains a separate function called generate_dictionary_script. This can be used to read in a modified dictionary and will then give out the code to insert into this function.
}
