% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{create_iis_to_sis_candidates}
\alias{create_iis_to_sis_candidates}
\title{Create IIS-to-SIS compression candidates}
\usage{
create_iis_to_sis_candidates(
  indicator_table,
  y_index,
  min_iis_run_length = 3,
  same_sign_ratio = 3
)
}
\description{
Create IIS-to-SIS compression candidates
}
\keyword{internal}
