% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{create_sis_to_iis_candidates}
\alias{create_sis_to_iis_candidates}
\title{Create SIS-to-IIS compression candidates}
\usage{
create_sis_to_iis_candidates(indicator_table, y_index, tolerance = 0.25)
}
\description{
Create SIS-to-IIS compression candidates
}
\keyword{internal}
