% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unit_root.R
\name{decide_unit_roots}
\alias{decide_unit_roots}
\title{Automated decision based on unit root tests.}
\usage{
decide_unit_roots(urtest, alpha = c("1pct", "5pct", "10pct"))
}
\arguments{
\item{urtest}{A named list as returned from \code{\link{test_unit_roots}}.}

\item{alpha}{Significance level for unit root tests.}
}
\value{
A named list as the input but with additional element
\code{$decision}.
}
\description{
Automated decision based on unit root tests.
}
