% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_codes.R
\name{determine_datacodes}
\alias{determine_datacodes}
\title{Determine which Eurostat datasets need to be downloaded}
\usage{
determine_datacodes(specification, dictionary = NULL)
}
\arguments{
\item{specification}{A tibble or data.frame as returned by
\code{\link{translate_variables}}.}

\item{dictionary}{A tibble or data.frame storing the Eurostat variable code
in column 'eurostat_code' and the Eurostat dataset id in column
'dataset_id' . When \code{NULL}, the \link[=dict]{default dictionary} is
used.}
}
\value{
Returns a list with two character vectors:
\describe{
\item{var.ids}{contains the Eurostat variable ids}
\item{data.ids}{contains the Eurostat dataset ids.}
}
}
\description{
Takes the translated specification, extracts the Eurostat variable codes, and
uses a dictionary to check from which Eurostat datasets they should be
retrieved.
}
