% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_codes.R
\name{determine_eurocodes}
\alias{determine_eurocodes}
\title{Determine which Eurostat variables need to be found}
\usage{
determine_eurocodes(specification, dictionary = NULL)
}
\arguments{
\item{specification}{A tibble or data.frame as returned by
\code{\link{translate_variables}}.}

\item{dictionary}{A tibble or data.frame storing the Eurostat variable code
in column 'eurostat_code'. When \code{NULL}, the
\link[=dict]{default dictionary} is used.}
}
\value{
Returns a list with one character vector:
\describe{
\item{var.ids}{contains the Eurostat variable ids}
}
}
\description{
Takes the translated specification and extracts the Eurostat variable codes
}
