% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_model.R
\name{diagnostics_model}
\alias{diagnostics_model}
\title{Creates a Summary of the Diagnostics of the Aggregate Model}
\usage{
diagnostics_model(model)
}
\arguments{
\item{model}{An aggregate model of class 'aggmod'}
}
\value{
Returns a data.frame with the p-values of the AR and ARCH
misspecification tests and the number of impulse and step indicators
retained for each module.
}
\description{
Creates a Summary of the Diagnostics of the Aggregate Model
}
