% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unit_root.R
\name{diagnostics_unit_root}
\alias{diagnostics_unit_root}
\title{Provide unit root diagnostics for the data.}
\usage{
diagnostics_unit_root(model)
}
\arguments{
\item{model}{An \code{\link[=new_osem]{osem}} object.}
}
\value{
A tibble storing the unit root test results for each variable that
appears in a modelled equation ("d"). Since log transoformations may differ
by whether the variable is a dependent or independent variable, separate
tests are conducted. THe tibble also records in which modules the variables
feature.
}
\description{
Provide unit root diagnostics for the data.
}
