% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{dict_edgar}
\alias{dict_edgar}
\title{Dictionary for translating EDGAR model variable names}
\format{
A tibble with 45 rows and 7 variables:
\describe{
\item{ipcc_sector}{EDGAR IPCC National Greenhouse Gas Inventories sector code,
see \url{https://www.ipcc-nggip.iges.or.jp/efdb/find_ef.php?reset=} and \url{https://www.ipcc-nggip.iges.or.jp/public/2006gl/index.html}}
\item{full_name}{Full name/description of the emission source}
\item{model_varname}{Variable name in the model equations, must be unique}
\item{dataset_id}{URL link to the dataset file on the web}
\item{database}{Name of the database, always "edgar" for this dictionary}
\item{geo}{ISO 3166-1 alpha-2 country code, e.g. "DE" for Germany}
\item{freq}{Frequency of the variable in the dataset, e.g. \code{"m"} for monthly}
}
}
\source{
EDGAR database, available at \url{https://edgar.jrc.ec.europa.eu/dataset_ghg80}.
}
\usage{
dict_edgar
}
\description{
A dataset containing EDGAR (Emissions Database for Global Atmospheric Research) model variable names
and all information required for obtaining emissions data. For details see \link{create_dictionaries}.
}
\details{
To see all details on IPCC sectors see \url{https://www.ipcc-nggip.iges.or.jp/efdb/find_ef.php?reset=} and \url{https://www.ipcc-nggip.iges.or.jp/public/2006gl/index.html}
}
\note{
Should new filters be necessary, users can simply add additional columns to the dictionary with names
that exactly match how they appear in the EDGAR data retrieval functions.
}
\keyword{datasets}
