% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{dict_eurostat}
\alias{dict_eurostat}
\title{Dictionary for translating Eurostat model variable names}
\format{
A tibble with 167 rows and 22 variables:
\describe{
\item{model_varname}{Variable name in the model equations, must be unique}
\item{full_name}{Full name/description of the variable}
\item{database}{Name of the database, always "eurostat" for this dictionary}
\item{variable_code}{Identifier of the variable, e.g. Eurostat variable code}
\item{dataset_id}{Identifier of the Eurostat dataset where the variable is available}
\item{var_col}{Name of variable/column in dataset}
\item{freq}{Frequency of the variable in the dataset. \code{"a"} for annual, \code{"q"} for quarterly, \code{"m"} for monthly}
\item{geo}{ISO 3166-1 alpha-2 country code, e.g. "DE" for Germany}
\item{unit}{Eurostat unit in which the variable is measured}
\item{s_adj}{Eurostat seasonal adjustment indicator}
\item{nace_r2}{Eurostat identifier for NACE Rev. 2 classification}
\item{cpa2_1}{Eurostat identifier for Classification of Products by Activity (CPA)}
\item{sector}{Sector classification where applicable}
\item{direct}{Direction indicator where applicable}
\item{age}{Age category where applicable}
\item{partner}{Partner country or entity where applicable}
\item{finpos}{Financial position indicator where applicable}
\item{p_adj}{Price adjustment indicator where applicable}
\item{meat}{Meat type classification where applicable}
\item{citizen}{Citizenship status where applicable}
\item{wstatus}{Work status where applicable}
\item{tra_oper}{Transport operation classification where applicable}
}
}
\source{
Eurostat database, available at \url{https://ec.europa.eu/eurostat}.
}
\usage{
dict_eurostat
}
\description{
A dataset containing Eurostat model variable names and all information required
for retrieving data from the Eurostat database. For details see \link{create_dictionaries}.
}
\note{
Should new filters be necessary, users can simply add additional columns to the dictionary with names
that exactly match how they appear in the eurostat R package. For example, if a eurostat query returns a column
with a specific name, that exact name should be used as a column in this dictionary.
}
\keyword{datasets}
