% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{dict_imf}
\alias{dict_imf}
\title{Dictionary for translating IMF model variable names}
\format{
A tibble with 1 row and 8 variables:
\describe{
\item{model_varname}{Unique variable name used in model equations}
\item{full_name}{Full description of the variable}
\item{database}{Name of the data source, here \code{"imf"}}
\item{dataset_id}{Identifier of the IMF dataset}
\item{freq}{Frequency of the data, with \code{"M"} for monthly}
\item{ref_area}{Geographical reference area, e.g., \code{"W00"} for world}
\item{commodity}{Type of commodity, such as \code{"POILAPSP"} for oil}
\item{unit_measure}{Unit of measurement, e.g., \code{"USD"} for U.S. dollars}
}
}
\source{
Data retrieved from IMF, available at \url{https://www.imf.org/en/Data}.
}
\usage{
dict_imf
}
\description{
A dataset containing IMF model variable names and metadata for data retrieval. For details see \link{create_dictionaries}.
}
\note{
Should new filters be necessary, users can simply add additional columns to the dictionary with names
that exactly match how they appear in the imf.data package. For example, if an imf.data query returns a column
with a specific name, that exact name should be used as a column in this dictionary.
}
\keyword{datasets}
