% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{dict_statcan}
\alias{dict_statcan}
\title{Statcan dictionary object}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
\item{geo}{Geographic location for which the data was downloaded.}
\item{na_item}{Variable Name according to the osem dictionary.}
\item{time}{The date of the observation.}
\item{s_adj}{Indication whether variable was seasonally adjusted or not. SCA refers to seasonal and calendar adjusted data.}
\item{nace_r2}{If applicable, the NACE2 sector code.}
\item{ipcc_sector}{EDGAR IPCC National Greenhouse Gas Inventories,
see \url{https://www.ipcc-nggip.iges.or.jp/public/2006gl/index.html}.
\code{"TOTAL"} is not an official IPCC code but is internally interpreted to use country totals.}
\item{cpa2_1}{Eurostat identifier for Classification of Products by Activity (CPA)}
North American Industry Classification System (NAICS)
North American Product Classification System (NAPCS)

}
}
\usage{
dict_statcan
}
\description{
Sample sample dataset that hs been downloaded using the functions in the osem package
}
\details{
Functions used to obtain data:
statCan_load_or_download, IMF_load_or_download, or download_eurostat
These functions will give outout consistent to download_eurostat (a function witin load_or_download_variables)

geo,na_item,values,time,s_adj,nace_r2
}
\keyword{datasets}
