% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_download_variables.R
\name{download_edgar}
\alias{download_edgar}
\title{Download Edgar}
\usage{
download_edgar(to_obtain, quiet)
}
\arguments{
\item{to_obtain}{A data.frame as returned by \code{\link[=determine_variables]{determine_variables()}}.}

\item{quiet}{Logical with default = \code{FALSE}. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}
}
\value{
Returns a list with two named elements: \code{$df} stores the
downloaded data and \code{$to_obtain} the updated data.frame tracking which
variables still need to be obtained.
}
\description{
Downloads Edgar data given a data.frame of required variables
}
