% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_download_variables.R
\name{download_eurostat}
\alias{download_eurostat}
\title{Download Eurostat}
\usage{
download_eurostat(to_obtain, additional_filters, quiet)
}
\arguments{
\item{to_obtain}{A data.frame as returned by \code{\link[=determine_variables]{determine_variables()}}.}

\item{additional_filters}{A character vector storing the column names of
additional filters if present. Length 0 if not present.}

\item{quiet}{Logical with default = \code{FALSE}. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}
}
\value{
Returns a list with two named elements: \code{$df} stores the
downloaded data and \code{$to_obtain} the updated data.frame tracking which
variables still need to be obtained.
}
\description{
Downloads Eurostat data given a data.frame of required variables
}
