% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{empty_candidate_tbl}
\alias{empty_candidate_tbl}
\title{Empty candidate tibble with stable columns}
\usage{
empty_candidate_tbl()
}
\description{
Empty candidate tibble with stable columns
}
\keyword{internal}
