% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{evaluate_indicator_candidates}
\alias{evaluate_indicator_candidates}
\title{Evaluate indicator compression candidates}
\usage{
evaluate_indicator_candidates(
  model_object,
  candidates,
  criterion = c("BIC", "AIC"),
  require_diagnostics_not_worse = FALSE,
  diagnostic_pval_floor = 0.01
)
}
\description{
Evaluate indicator compression candidates
}
\keyword{internal}
