% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{extract_indicator_table}
\alias{extract_indicator_table}
\title{Extract retained IIS/SIS indicators from a gets/isat object}
\usage{
extract_indicator_table(model_object)
}
\description{
Extract retained IIS/SIS indicators from a gets/isat object
}
\keyword{internal}
