% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_build_draw_term_data.R
\name{forecast_build_draw_term_data}
\alias{forecast_build_draw_term_data}
\title{Build forecast regressors for every upstream uncertainty path}
\usage{
forecast_build_draw_term_data(
  state_data,
  deterministic_data,
  deterministic_draw_data,
  recipe,
  n.ahead,
  uncertainty_sample
)
}
\description{
Build forecast regressors for every upstream uncertainty path
}
\keyword{internal}
