% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_cvar.R
\name{forecast_cvar}
\alias{forecast_cvar}
\title{Forecast CVAR Model (internal)}
\usage{
forecast_cvar(
  model,
  i,
  exog_df_ready,
  full_exog_predicted_data,
  n.ahead,
  current_spec,
  prediction_list,
  uncertainty_sample,
  nowcasted_data
)
}
\arguments{
\item{model}{The overall 'osem' model as returned by \code{\link[osem]{run_model}}}

\item{i}{Index of the CVAR model within the module collection}

\item{exog_df_ready}{Data frame of exogenous variables prepared for forecasting}

\item{full_exog_predicted_data}{Full data frame of predicted exogenous variables}

\item{n.ahead}{Number of steps ahead to forecast}

\item{current_spec}{Current module specification}

\item{prediction_list}{List of predictions from previous modules}

\item{uncertainty_sample}{Number of uncertainty samples to generate}

\item{nowcasted_data}{Nowcasted data for the current module}
}
\value{
A list with class 'osem.forecast.module' containing the central and all estimates forecasts.
}
\description{
This internal function generates forecasts from a CVAR (Cointegrated Vector Autoregression) model within the OSEM framework.
}
