% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_failure.R
\name{forecast_failure}
\alias{forecast_failure}
\title{Determine the number of Forecast Failures in Insample Forecasting}
\usage{
forecast_failure(forecast, data)
}
\arguments{
\item{forecast}{A forecast object as returned by \code{\link{forecast_model}}.}

\item{data}{A tibble or data.frame containing the original data used for estimation.}
}
\value{
A tibble representing information about the success and failures of forecasting.
}
\description{
Determine the number of Forecast Failures in Insample Forecasting
}
\examples{
\donttest{
specification <- dplyr::tibble(
 type = c(
   "n"
 ),
 dependent = c(
   "FinConsExpHH"
 ),
 independent = c(
   "FinConsExpGov + HICP_Gas"
 )
)

set.seed(123)
testdata <- dplyr::tibble(
 time = seq.Date(from = as.Date("2005-01-01"),
                 to = as.Date("2023-10-01"),
                 by = "quarter"),
 FinConsExpGov = rnorm(mean = 100, n = length(time)),
 HICP_Gas = rnorm(mean = 200, n = length(time)),
 FinConsExpHH  = 0.5 + 0.2*FinConsExpGov + 0.3 *
   HICP_Gas + rnorm(length(time), mean = 0, sd = 0.2))

testdata <- tidyr::pivot_longer(testdata,
                               cols = -time,
                               names_to = "na_item",
                               values_to = "values")

model <- run_model(specification = specification,
                  dictionary = dict,
                  input = testdata,
                  primary_source = "local",
                  present = FALSE,
                  quiet = TRUE,
                  saturation = "IIS")

insample_output <- forecast_insample(model, sample_share = 0.97)
insample_output$forecast_failures
}
}
