% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_identities.R
\name{forecast_identities}
\alias{forecast_identities}
\title{An internal function to forecast identities within forecast_model}
\usage{
forecast_identities(
  model,
  exog_df_ready,
  current_spec,
  prediction_list,
  uncertainty_sample
)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}

\item{uncertainty_sample}{Integer. Number of draws to be made for the error bars. Default is 100.}
}
\value{
A list that is then used by forecast_model to set-up the final prediction obejct. The list contains the identity estimate.
}
\description{
An internal function to forecast identities within forecast_model
}
