% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_module_identity.R
\name{forecast_module_identity}
\alias{forecast_module_identity}
\title{An internal function to forecast identities within forecast_model}
\usage{
forecast_module_identity(
  model,
  exog_df_ready,
  current_spec,
  prediction_list,
  uncertainty_sample
)
}
\arguments{
\item{model}{A model object of class 'osem'.}

\item{exog_df_ready}{The outcome of \link[=forecast_exogenous_values]{forecast_exogenous_values} function (potentially modified by the \link[=nowcasting]{nowcasting} function)}

\item{current_spec}{Current specification. Is defined within \link[=forecast_model]{forecast_model}.}

\item{prediction_list}{Current List containing all predictions. Is defined within \link[=forecast_model]{forecast_model}.}

\item{uncertainty_sample}{Integer. Number of draws to be made for the error bars. Default is 100.}
}
\value{
A list that is then used by forecast_model to set-up the final prediction object. The list contains the identity estimate.
}
\description{
An internal function to forecast identities within forecast_model
}
