% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_recipe_compile.R
\name{forecast_recipe_compile}
\alias{forecast_recipe_compile}
\title{Compile the information required to forecast an estimated OSEM module}
\usage{
forecast_recipe_compile(
  model_object,
  model_form,
  dep_var_basename,
  x_vars_basename,
  use_logs,
  transformations,
  term_specs,
  lag_only_vars = character(0)
)
}
\description{
Compile the information required to forecast an estimated OSEM module
}
\keyword{internal}
