% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_setup_estimated_relationships.R
\name{forecast_setup_estimated_relationships}
\alias{forecast_setup_estimated_relationships}
\title{Internal function to set-up the forecasting of estimated relationships}
\usage{
forecast_setup_estimated_relationships(
  model,
  i,
  exog_df_ready,
  n.ahead,
  current_spec,
  prediction_list,
  uncertainty_sample,
  nowcasted_data = NULL,
  full_exog_predicted_data = NULL
)
}
\arguments{
\item{model}{A model object of class 'osem'.}

\item{i}{Current module that is being cycled through}

\item{exog_df_ready}{Outcome of forecast_exogenous_values() which is the set of forecasted exogenous values}

\item{n.ahead}{Periods to forecast ahead}

\item{current_spec}{The current specification for the module being forecasted}

\item{prediction_list}{The full list of all predictions. The results of the function will be saved in this list.}

\item{uncertainty_sample}{Integer. Number of draws to be made for the error bars. Default is 100.}

\item{nowcasted_data}{The full_data element of the model object resulting from the nowcasting() function. Used to substitute missing historical data.}

\item{full_exog_predicted_data}{An argument to pass a larger data.frame to the function that can contain the entire exogenously predicted data. This is an argument that is needed in the nowcasting() function.}
}
\value{
A list containing, among other elements, the data required to carry out the forecast for this estimated module.
The elements contain:
\itemize{
\item current_pred_raw: central future source data, before all required differences/lags are constructed.
\item current_pred_raw_all: the same source data, with upstream uncertainty stored in list-columns.
\item pred_df: complete prediction-ready regressor data for the central forecast.
\item pred_df.all: a list of complete prediction-ready regressor data frames, one per uncertainty run.
}
}
\description{
Internal function to set-up the forecasting of estimated relationships
}
