% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_module_data.R
\name{identify_module_data}
\alias{identify_module_data}
\title{Identifies which variables are required for running the module}
\usage{
identify_module_data(module, classification, data)
}
\arguments{
\item{module}{A row of the specification table.}

\item{classification}{A data.frame with two columns: 'var' stores the
Eurostat variable code and 'class' the classification code (one of
\code{"n"}, \code{"d"}, or \code{"x"}) reflecting whether the variable is
endogenous by modelling, endogenous by identity/definition, or exogenous.
As returned by \link{classify_variables}.}

\item{data}{A dataset containing at least the required dependent variable and
independent variables for the module.}
}
\value{
Returns a subset of the original \code{data} that only contains the
relevant variables for the module.
}
\description{
Identifies the required variables for running the module and extracts that
subset.
}
