% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_module.R
\name{identity_module}
\alias{identity_module}
\title{Performs computations for the identity/definition modules}
\usage{
identity_module(module, data, classification)
}
\arguments{
\item{module}{A row of the specification table.}

\item{data}{A dataset containing at least the required dependent variable and
independent variables for the module.}

\item{classification}{A data.frame with two columns: 'var' stores the
Eurostat variable code and 'class' the classification code (one of
\code{"n"}, \code{"d"}, or \code{"x"}) reflecting whether the variable is
endogenous by modelling, endogenous by identity/definition, or exogenous.
As returned by \link{classify_variables}.}
}
\value{
Returns the original tibble or data.frame \code{data} in wide format
and appended with the fitted values of the module.
}
\description{
Calculates the fitted values for identity/definition modules using previously
fitted values of endogenous variables.
}
