% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{inverse_transform_osem_values}
\alias{inverse_transform_osem_values}
\title{Invert an OSEM variable transformation}
\usage{
inverse_transform_osem_values(x, transformation = "none")
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{transformation}{Character. One of \code{"none"}, \code{"log"}, or \code{"asinh"}.}
}
\value{
Values on the original level scale.
}
\description{
Invert an OSEM variable transformation
}
\keyword{internal}
