% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_download_variables.R
\name{load_locally}
\alias{load_locally}
\title{Load Locally}
\usage{
load_locally(to_obtain, input, quiet)
}
\arguments{
\item{to_obtain}{A data.frame as returned by \code{\link[=determine_variables]{determine_variables()}}.}

\item{input}{A character vector of file paths, a data.frame, or a list of
data.frames or file paths. Required if \code{download == FALSE} and at least
one variable is specified as coming from a local source.}

\item{quiet}{Logical with default = \code{FALSE}. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}
}
\value{
Returns a list with two named elements: \code{$df} stores the
downloaded data and \code{$to_obtain} the updated data.frame tracking which
variables still need to be obtained.
}
\description{
Loads data from a local directory
}
