% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_indicators.R
\name{make_indicator_column}
\alias{make_indicator_column}
\title{Construct IIS/SIS indicator column from term name and model index}
\usage{
make_indicator_column(term, y_index)
}
\description{
Construct IIS/SIS indicator column from term name and model index
}
\keyword{internal}
