% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\name{model_table}
\alias{model_table}
\title{OSEM Model Table}
\usage{
model_table(model, output = "default", title = "OSEM Model Results")
}
\arguments{
\item{model}{An OSEM model object or class \code{osem}, created by the \code{run_model} function.}

\item{output}{Character. Either "default", "html", "data.frame", or "latex". Default is "default" (which is the default in the \code{\link[modelsummary]{modelsummary}} function.}

\item{title}{Character. The title of the table. Default is "OSEM Model Results".}
}
\value{
A table summarizing the regression results of the OSEM model (either an html or a text (latex) table).
}
\description{
Show the final regression equations estimated by the OSEM Model
}
\examples{
spec <- dplyr::tibble(
  type = c(
    "d",
    "d",
    "n"
  ),
  dependent = c(
    "StatDiscrep",
    "TOTS",
    "Import"
  ),
  independent = c(
    "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
    "GValueAdd + Import",
    "FinConsExpHH + GCapitalForm"
  )
)
\donttest{
a <- run_model(specification = spec,
              input = sample_input,
              primary_source = "local",
              constrain.to.minimum.sample = FALSE)
model_table(a)
}

}
