% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Visualize dependence between modules}
\usage{
network(model, layout = "kk")
}
\arguments{
\item{model}{A model of class 'osem'}

\item{layout}{Character. The layout of the network graph as per the \code{ggraph} package. Default is "kk", try "auto", "fr" (Fruchterman-Reingold), or "dh" (Davidson-Harel).}
}
\value{
Returns a network graph illustrating the dependence between the
different modules.
}
\description{
Visualize dependence between modules
}
