% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_aggmod}
\alias{new_aggmod}
\title{Simple constructor for class 'aggmod'}
\usage{
new_aggmod(x = list())
}
\arguments{
\item{x}{A list storing the aggregate model output.}
}
\value{
An object of class \code{"aggmod"}, which is a named list with four elements:
\describe{
\item{args}{A named list storing the user arguments for the aggregate
model.}
\item{module_order}{The original but ordered specification.}
\item{module_collection}{The above specification with two added columns
that store the model object for each module and the dataset used for
estimation, including fitted values for the dependent variable.}
\item{full_data}{A tibble or data.frame containing the complete original
data for the aggregate model and the fitted values of each module.}
}
}
\description{
Simple constructor for class 'aggmod'
}
\keyword{internal}
