% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_osem}
\alias{new_osem}
\title{Simple constructor for class 'osem'}
\usage{
new_osem(x = list())
}
\arguments{
\item{x}{A list storing the osem output.}
}
\value{
An object of class \code{"osem"}, which is a named list with seven elements:
\describe{
\item{args}{A named list storing the user arguments for the OSEM
model.}
\item{module_order}{The original but ordered specification.}
\item{module_collection}{The above specification with two added columns
that store the model object for each module and the dataset used for
estimation, including fitted values for the dependent variable.}
\item{processed_input_data}{A tibble or data.frame containing the processed
input data, including variables that are constructed by definition.}
\item{full_data}{A tibble or data.frame containing the complete original
data for the OSEM model and the fitted values of each module.}
\item{dictionary}{A tibble storing information about each variable,
including from where it is obtained.}
\item{opts_df}{A tibble storing for each module whether the involved
variables have been transformed using the log or asinh function.}
}
}
\description{
Simple constructor for class 'osem'
}
\keyword{internal}
