% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.model.R
\docType{package}
\name{osem}
\alias{osem}
\title{\code{osem} package}
\description{
The Package for the Open Source Empirical Macro (OSEM) Model
}
\details{
See the README on
\href{https://cran.r-project.org/package=osem/README.html}{CRAN}
or \href{https://github.com/moritzpschwarz/osem#readme}{GitHub}
}
