% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aggmod.forecast.R
\name{plot.aggmod.forecast}
\alias{plot.aggmod.forecast}
\title{Plot a Forecast Object of the Aggregate Model}
\usage{
\method{plot}{aggmod.forecast}(
  x,
  exclude.exogenous = TRUE,
  order.as.run = FALSE,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class aggmod.forecast, which is the output from the forecast_model function.}

\item{exclude.exogenous}{Logical. Should exogenous values be plotted? Default is FALSE.}

\item{order.as.run}{Logical. Should the plots be arranged in the way that the model was run? Default FALSE.}

\item{...}{Further arguments (currently not in use).}
}
\description{
Plot a Forecast Object of the Aggregate Model
}
\examples{
spec <- dplyr::tibble(
type = c(
  "d",
  "d",
  "n"
),
dependent = c(
  "StatDiscrep",
  "TOTS",
  "Import"
),
independent = c(
  "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
  "GValueAdd + Import",
  "FinConsExpHH + GCapitalForm"
)
)

fa <- list(geo = "AT", s_adj = "SCA", unit = "CLV05_MEUR")
fb <- list(geo = "AT", s_adj = "SCA", unit = "CP_MEUR")
filter_list <- list("P7" = fa, "YA0" = fb, "P31_S14_S15" = fa,
"P5G" = fa, "B1G" = fa, "P3_S13" = fa, "P6" = fa)
\donttest{
a <- run_model(specification = spec, dictionary = NULL,
inputdata_directory = NULL, filter_list = filter_list, download = TRUE,
save_to_disk = NULL, present = FALSE)
plot(forecast_model(a))
}
}
