% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_model.R
\name{present_model}
\alias{present_model}
\title{Present the results of the OSEM Model in a Shiny Dashboard}
\usage{
present_model(model)
}
\arguments{
\item{model}{object. The output of the run_model command.}
}
\value{
Opens a shiny app.
}
\description{
Present the results of the OSEM Model in a Shiny Dashboard
}
