% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aggmod.R
\name{print.aggmod}
\alias{print.aggmod}
\title{Print output of an Aggregate Model}
\usage{
\method{print}{aggmod}(x, ...)
}
\arguments{
\item{x}{An object of type 'aggmod'}

\item{...}{Further arguments.}
}
\description{
Print output of an Aggregate Model
}
