% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.osem.forecast.insample.R
\name{print.osem.forecast.insample}
\alias{print.osem.forecast.insample}
\title{Print output of an OSEM Insample Forecast}
\usage{
\method{print}{osem.forecast.insample}(x, plot = TRUE, full_names = FALSE, ...)
}
\arguments{
\item{x}{An object of class osem.forecast.insample, which is the output from the \link{forecast_insample} function.}

\item{plot}{Logical. Default = \code{TRUE}. Should the osem insample forecast output be plotted?}

\item{full_names}{Logical. Default = \code{FALSE}. Should the full names of the variables from the dictionary be printed?}

\item{...}{Further arguments.}
}
\description{
Print output of an OSEM Insample Forecast
}
