% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.osem.forecast.R
\name{print.osem.forecast}
\alias{print.osem.forecast}
\title{Print output of an OSEM Forecast}
\usage{
\method{print}{osem.forecast}(x, plot = TRUE, full_names = FALSE, ...)
}
\arguments{
\item{x}{An object of type 'osem.forecast'}

\item{plot}{Logical. Default = \code{TRUE}. Should the osem forecast output be plotted?}

\item{full_names}{Logical. Default = \code{FALSE}. Should the full names of the variables from the dictionary be printed?}

\item{...}{Further arguments.}
}
\value{
A printed summary of the OSEM forecast output, including the forecast horizon, method, and central estimates.
}
\description{
Print output of an OSEM Forecast
}
