% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isat.R
\name{run_isat}
\alias{run_isat}
\title{Internal function to run and prepare isat estimation}
\usage{
run_isat(
  yvar,
  y.name,
  xvars,
  clean_data,
  ar = NULL,
  saturation,
  saturation.tpval,
  max.block.size,
  determine.blocksize
)
}
\arguments{
\item{yvar}{The y variable}

\item{y.name}{The name of the y variable.}

\item{xvars}{All x variables.}

\item{clean_data}{An input data.frame or tibble. Must be the output of clean_data() to fit all requirements.}

\item{ar}{numeric vector or \code{NULL}. The AR order to be used.}

\item{saturation}{Carry out Indicator Saturation using the 'isat' function in the 'gets' package. Needs a character vector or string. Default is 'c("IIS","SIS")' to carry out Impulse Indicator Saturation and Step Indicator Saturation. Other possible values are 'NULL' to disable or 'TIS' or Trend Indicator Saturation. When disabled, estimation will be carried out using the 'arx' function from the 'gets' package.}

\item{saturation.tpval}{The target p-value of the saturation methods (e.g. SIS and IIS, see the 'isat' function in the 'gets' package). Default is 0.01.}

\item{max.block.size}{Integer. Maximum size of block of variables to be selected over, default = 20.}

\item{determine.blocksize}{Logical. Whether max.block.size or needs to be determined.}
}
\value{
An isat object.
}
\description{
Internal function to run and prepare isat estimation
}
