% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_configuration.R
\name{run_shiny_configuration}
\alias{run_shiny_configuration}
\title{Creates and runs the Shiny Configuration App "shinyconfigaggmodel"}
\usage{
run_shiny_configuration(model = NULL)
}
\arguments{
\item{model}{The model object that is passed by the 'present_model()' function.}
}
\description{
Opens a window for the user to interact with the app. The user can upload
an \link[=new_aggmod]{aggmod} object returned by \code{\link{run_model}} and
produce graphical and tabular output.
}
