% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{sample_input}
\alias{sample_input}
\title{Sample Input for the Aggregate Model}
\format{
A tibble with 1200 rows and 9 variables:
\describe{
\item{unit}{Unit according to Eurostat.}
\item{s_adj}{Indication whether variable was seasonally adjusted or not. SCA refers to seasonal and calendar adjusted data.}
\item{na_item}{Variable Name according to the aggregate.model dictionary.}
\item{geo}{Geographic location for which the data was downloaded.}
\item{time}{The date of the observation.}
\item{nace_r2}{If applicable, the NACE2 sector code.}
\item{p_adj}{If applicable, price adjustment.}
\item{cpa2_1}{If applicable, Classification of Products by Activity.}
}
}
\source{
Own compilation, data from \url{https://ec.europa.eu/eurostat}, downloaded with aggregate.model::run_model()
}
\usage{
sample_input
}
\description{
A dataset containing input data that has been downloaded using the functions in the aggregate.model package.
This dataset is supposed to be used for testing and to enable out-of-the-box estimation.
}
\keyword{datasets}
