% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{statcan_dict}
\alias{statcan_dict}
\title{Statcan dictionary object}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
\item{Col1}{description of Col1}
\item{Col2}{description of Col2}
}
}
\usage{
statcan_dict
}
\description{
Description_of_the_Data
}
\keyword{datasets}
