% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{strsplits}
\alias{strsplits}
\title{unpack formulas}
\usage{
strsplits(x, splits, ...)
}
\arguments{
\item{x}{String to split}

\item{splits}{vector of elements to split the string by.}

\item{...}{Further arguments.}
}
\description{
NOTE: exported for now for interactive programming, later will be internal
}
