% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{strsplits2}
\alias{strsplits2}
\title{unpack formulas v2}
\usage{
strsplits2(x, splits, ...)
}
\arguments{
\item{x}{String to split}

\item{splits}{vector of elements to split the string by.}

\item{...}{Further arguments.}
}
\description{
NOTE: adapted from strsplits to always return a character, not NULL
}
