% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unit_root.R
\name{test_unit_roots}
\alias{test_unit_roots}
\title{Run three types of ADF unit root tests.}
\usage{
test_unit_roots(x, max.ar, selectlags = c("Fixed", "AIC", "BIC"))
}
\arguments{
\item{x}{A numeric vector.}

\item{max.ar}{Maximum lag for augmented Dickey-Fuller test.}

\item{selectlags}{Character specifying "Fixed" lags or automatic lag
selection with information criteria "AIC" or "BIC".}
}
\value{
A named list with four elements, storing the function arguments and
the resulting three ADF unit root tests: (i) no deterministic terms, (ii) an
intercept, (iii) intercept and trend.
}
\description{
Run three types of ADF unit root tests.
}
