% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_variables.R
\name{translate_variables}
\alias{translate_variables}
\title{Translate model variable names to Eurostat codes}
\usage{
translate_variables(specification, dictionary = NULL)
}
\arguments{
\item{specification}{A tibble or data.frame with one column named 'dependent'
that contains the LHS (Y) variables and one column named 'independent' that
contains the RHS (x) variables, separated by + and - . Column 'type' should
specify whether the dependent variable is endogenous by modelling
(\code{"n"}) or endogenous by accounting definition (\code{"d"}).}

\item{dictionary}{A tibble or data.frame storing the model variable names in
a column named 'model_varname' and the corresponding Eurostat variable code
in 'eurostat_code'. When \code{NULL}, the \link[=dict]{default dictionary}
is used.}
}
\value{
Returns a tibble or data.frame as the input \code{specification} with
two added columns for the model variable name (\code{model_varname}) and
the dataset identifier (\code{dataset_id}).
}
\description{
Takes a specification and translates the model variable names into Eurostat
codes. Also adds the Eurostat dataset identifier from which the variable is
extracted.
}
