% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{update_data}
\alias{update_data}
\title{Updates the osem dataset with fitted values}
\usage{
update_data(orig_data, new_data)
}
\arguments{
\item{orig_data}{Original data that is to be updated.}

\item{new_data}{New dataset that will update the original data.}
}
\description{
NOTE: exported for now, later will be internal
}
