% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_original_data.R
\name{add_to_original_data}
\alias{add_to_original_data}
\title{Add the estimated fitted values back to the original}
\usage{
add_to_original_data(
  clean_data,
  model_object,
  dep_var_basename = "imports_of_goods_and_services",
  model_type = "ardl",
  opts_df,
  module
)
}
\arguments{
\item{clean_data}{An input data.frame or tibble. Must be the output of
clean_data() to fit all requirements.}

\item{model_object}{Most likely should be returned by the 'estimate_module()'
function.}

\item{dep_var_basename}{A character string of the name of the dependent
variable as contained in clean_data() in a level form (i.e. no ln or D in
front of the name).}

\item{model_type}{Either 'ardl', 'ecm', or 'cvar'.}

\item{opts_df}{Internal object containing detailed options and information on individual modules.}

\item{module}{A row of the specification table.}
}
\value{
A tibble with the fitted values as one column.
}
\description{
Add the estimated fitted values back to the original
}
