% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.model.R
\docType{package}
\name{aggregate.model}
\alias{aggregate.model}
\title{\code{aggregate.model} package}
\description{
The Package for the Aggregate Model
}
\details{
See the README on
\href{https://cran.r-project.org/package=aggregate.model/README.html}{CRAN}
or \href{https://github.com/moritzpschwarz/aggregate.model#readme}{GitHub}
}
