% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_table.R
\name{check_config_table}
\alias{check_config_table}
\title{Check the configuration of the model contained in the config table}
\usage{
check_config_table(config_table)
}
\arguments{
\item{config_table}{A tibble or data.frame with one column named 'dependent', containing the LHS (Y variables) and one named 'independent' containing the RHS (x variables separated by + and - ).}
}
\value{
A tibble that gives the order of the modules to be run.
}
\description{
Check the configuration of the model contained in the config table
}
\examples{
config_table_small <- dplyr::tibble(
  type = c("d","d","n"),
  dependent = c("JL", "TOTS", "B"),
  independent = c("TOTS - CP - CO - J - A", "YF + B", "CP + J")
)
aggregate.model:::check_config_table(config_table_small)

}
