% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Prepare the clean dataframe for modelling}
\usage{
clean_data(raw_data, max.ar = 4, max.dl = 2, trend = TRUE)
}
\arguments{
\item{raw_data}{A tibble or data.frame with the y variable and the x
variables. Needs to have a column called 'time', which is of class
\code{\link[base:Dates]{Date}}. Variable names need to be in column
'na_item', and values in column 'values'.}

\item{max.ar}{Integer. The maximum number of lags to use for the AR terms. as well as for the independent variables.}

\item{max.dl}{Integer. The maximum number of lags to use for the independent variables (the distributed lags).}

\item{trend}{Logical. Should a trend be added? Default is TRUE.}
}
\value{
A tibble with the cleaned data.
}
\description{
Prepare the clean dataframe for modelling
}
\examples{
sample_data <- dplyr::tibble(
  time = rep(seq.Date(
    from = as.Date("2000-01-01"),
    to = as.Date("2000-12-31"), by = 1
  ), each = 2),
  na_item = rep(c("yvar", "xvar"), 366), values = rnorm(366 * 2, mean = 100)
)
aggregate.model:::clean_data(sample_data, max.lag = 4)
}
