% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_function.R
\name{clean_data}
\alias{clean_data}
\title{Prepare the clean dataframe for modelling}
\usage{
clean_data(raw_data, preestimated_xvars = NULL, max.lag = 4)
}
\arguments{
\item{raw_data}{A tibble or data.frame with the y variable and the x variables. Needs to have a column called "time", which is off class = date. Variable names need to be in 'na_item', values in 'values'.}

\item{preestimated_xvars}{A data.frame or tibble with the x variables that have been pre-estimated by another equation in levels. Needs to have a column called "time", which is off class = date.}
}
\value{
A tibble with the cleaned data.
}
\description{
Prepare the clean dataframe for modelling
}
\examples{
sample_data <- tibble(time = rep(seq.Date(from = as.Date("2000-01-01"), to = as.Date("2000-12-31"),by = 1),each = 2), na_item = rep(c("yvar","xvar"),366), values = rnorm(366*2,mean = 100))
clean_data(sample_data, max.lag = 4)

}
