% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_codes.R
\name{determine_variables}
\alias{determine_variables}
\title{Determine which model variables need to be found}
\usage{
determine_variables(specification, dictionary)
}
\description{
Takes the specification and returns a dictionary-like data frame with the
required model variables, their filters, and where they can be found.
}
