% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dictionary for translating model variable names to Eurostat codes}
\format{
A tibble with 13 rows and 4 variables:
\describe{
\item{eurostat_code}{Eurostat identifier of the variable}
\item{model_varname}{Variable name in the model equations}
\item{full_name}{Full name of the variable according to Eurostat}
\item{dataset_id}{Eurostat identifier for the dataset where the variable is available}
\item{var_col}{Name of variable/column in Eurostat dataset}
\item{nace_r2}{Eurostat identifier for NACE Rev. 2 classification}
\item{cpa2_1}{Eurostat identifier for Classification of Products by Activity (CPA)}
}
}
\source{
Own compilation, codes from \url{https://ec.europa.eu/eurostat}
}
\usage{
dict
}
\description{
A dataset containing the model variable names, Eurostat code, full name, and
dataset id where the variable can be found.
}
\keyword{datasets}
