% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_module.R
\name{estimate_module}
\alias{estimate_module}
\title{Estimate the specific module using indicator saturation}
\usage{
estimate_module(
  clean_data,
  dep_var_basename,
  x_vars_basename,
  use_logs = "both",
  trend = TRUE,
  ardl_or_ecm = "ardl",
  max.ar = 4,
  max.dl = 2,
  saturation = c("IIS", "SIS"),
  saturation.tpval = 0.01,
  max.block.size = 20,
  gets_selection = TRUE,
  selection.tpval = 0.01,
  quiet = FALSE
)
}
\arguments{
\item{clean_data}{An input data.frame or tibble. Must be the output of clean_data() to fit all requirements.}

\item{dep_var_basename}{A character string of the name of the dependent variable as contained in clean_data() in a level form (i.e. no ln or D in front of the name).}

\item{x_vars_basename}{A character vector of the name(s) of the independent variable(s) as contained in clean_data() in a level form (i.e. no ln or D in front of the name).}

\item{use_logs}{To decide whether to log any variables. Must be one of 'both', 'y', 'x', or 'none'. Default is 'both'.}

\item{trend}{Logical. To determine whether a trend should be added. Default is TRUE.}

\item{ardl_or_ecm}{Either 'ardl' or 'ecm' to determine whether to estimate the model as an Autoregressive Distributed Lag Function (ardl) or as an Equilibrium Correction Model (ecm).}

\item{max.ar}{Integer. The maximum number of lags to use for the AR terms. as well as for the independent variables.}

\item{max.dl}{Integer. The maximum number of lags to use for the independent variables (the distributed lags).}

\item{saturation}{Carry out Indicator Saturation using the 'isat' function in the 'gets' package. Needs a character vector or string. Default is 'c("IIS","SIS")' to carry out Impulse Indicator Saturation and Step Indicator Saturation. Other possible values are 'NULL' to disable or 'TIS' or Trend Indicator Saturation. When disabled, estimation will be carried out using the 'arx' function from the 'gets' package.}

\item{saturation.tpval}{The target p-value of the saturation methods (e.g. SIS and IIS, see the 'isat' function in the 'gets' package). Default is 0.01.}

\item{max.block.size}{Integer. Maximum size of block of variables to be selected over, default = 20.}

\item{gets_selection}{Logical. Whether general-to-specific selection using the 'getsm' function from the 'gets' package should be done on the final saturation model. Default is TRUE.}

\item{selection.tpval}{Numeric. The target p-value of the model selection methods (i.e. general-to-specific modelling, see the 'getsm' function in the 'gets' package). Default is 0.01.}

\item{quiet}{Logical. Should messages about the forecast procedure be suppressed?}
}
\value{
A list containing all estimated models, with the model with the smallest BIC under 'best_model'.
}
\description{
Estimate the specific module using indicator saturation
}
\examples{
sample_data <- dplyr::tibble(
  time = rep(seq.Date(
    from = as.Date("2000-01-01"),
    to = as.Date("2000-12-31"), by = 1
  ), each = 2),
  na_item = rep(c("yvar", "xvar"), 366), values = rnorm(366 * 2, mean = 100)
)
sample_data_clean <- aggregate.model:::clean_data(sample_data, max.ar = 4)
aggregate.model:::estimate_module(sample_data_clean, "yvar", "xvar")

}
