% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_build_term_data.R
\name{forecast_build_term_data}
\alias{forecast_build_term_data}
\title{Build forecast regressors from a stored module recipe}
\usage{
forecast_build_term_data(state_data, deterministic_data, recipe)
}
\description{
Build forecast regressors from a stored module recipe
}
\keyword{internal}
