% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_insample.R
\name{forecast_insample}
\alias{forecast_insample}
\title{Forecast model insample to evaluate performance}
\usage{
forecast_insample(model, seed = 1234)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}

\item{seed}{Integer. Seed for the uncertainty draws to be made. Default is 1234.}
}
\value{
List element.
}
\description{
Forecast model insample to evaluate performance
}
\examples{
forecast_insample(model)
}
