% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_insample.R
\name{forecast_insample}
\alias{forecast_insample}
\title{Forecast model insample to evaluate forecasting performance}
\usage{
forecast_insample(
  model,
  sample_share = 0.5,
  seed = 1234,
  uncertainty_sample = 100,
  exog_fill_method = "AR",
  plot.forecast = TRUE
)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}

\item{sample_share}{Share of the sample that should be used for the insample forecasting. Must be a numeric and must be either 1 or smaller but larger than 0.}

\item{seed}{Integer. Seed for the uncertainty draws to be made. Default is 1234.}

\item{uncertainty_sample}{Integer. Number of draws to be made for the error bars. Default is 100.}

\item{exog_fill_method}{Character, either 'AR' or 'last'. When no exogenous values have been provided, these must be inferred. When option 'exog_fill_method = "AR"' then an autoregressive model is used to further forecast the exogenous values. With 'last', simply the last available value is used.}

\item{plot.forecast}{Logical. Should the result be plotted? Default is TRUE.}
}
\value{
List element.
}
\description{
Forecast model insample to evaluate forecasting performance
}
