% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_isat.R
\name{forecast_isat}
\alias{forecast_isat}
\title{Forecast ISAT Module (internal)}
\usage{
forecast_isat(
  model,
  i,
  exog_df_ready,
  exog_df_ready_full,
  n.ahead,
  current_spec,
  prediction_list,
  uncertainty_sample,
  ci.levels,
  nowcasted
)
}
\arguments{
\item{model}{The overall 'osem' model as returned by \code{\link[osem]{run_model}}}

\item{i}{The index of the current module within the model's module collection}

\item{exog_df_ready}{The exogenous data frame prepared for forecasting}

\item{exog_df_ready_full}{The full exogenous data frame prepared for forecasting}

\item{n.ahead}{Number of steps ahead to forecast}

\item{current_spec}{The current specification for the module being forecasted}

\item{prediction_list}{The collection of all predictions}

\item{uncertainty_sample}{The number of uncertainty samples to draw for the prediction}

\item{ci.levels}{The confidence interval levels for the prediction}

\item{nowcasted}{The nowcasted data for the model}
}
\value{
A tibble containing the updated prediction_list object with forecasts for the current module
}
\description{
This internal function generates forecasts from an ISAT (Indicator Saturation) model within the OSEM framework (see \code{\link[gets]{isat}}).
}
