% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_model.R
\name{forecast_model}
\alias{forecast_model}
\title{Forecast OSEM model}
\usage{
forecast_model(
  model,
  exog_predictions = NULL,
  n.ahead = 10,
  ci.levels = c(0.5, 0.66, 0.95),
  exog_fill_method = "AR",
  ar.fill.max = 4,
  plot = TRUE,
  uncertainty_sample = 100,
  quiet = FALSE
)
}
\arguments{
\item{model}{A model object of class 'osem'.}

\item{exog_predictions}{A data.frame or tibble with values for the exogenous values. The number of rows of this data must be equal to n.ahead.}

\item{n.ahead}{Periods to forecast ahead}

\item{ci.levels}{Numeric vector. Vector with confidence intervals to be calculated. Default: c(0.5,0.66,0.95)}

\item{exog_fill_method}{Character, either 'AR', 'auto', or 'last'. When no exogenous values have been provided, these must be inferred. When option 'exog_fill_method = "AR"' then an autoregressive model is used to further forecast the exogenous values. With 'last', simply the last available value is used. 'auto' is an \code{\link[forecast]{auto.arima}} model.}

\item{ar.fill.max}{Integer. When no exogenous values have been provided, these must be inferred. If option 'exog_fill_method = "AR"' then an autoregressive model is used to further forecast the exogenous values. This options determines the number of AR terms that should be used. Default is 4.}

\item{plot}{Logical. Should the result be plotted? Default is TRUE.}

\item{uncertainty_sample}{Integer. Number of draws to be made for the error bars. Default is 100.}

\item{quiet}{Logical. Should messages about the forecast procedure be suppressed?}
}
\value{
A list of class 'osem.forecast' with the following elements:
\describe{
\item{forecast}{A tibble with the forecasted values for each module.}
\item{orig_model}{The original model object of class 'osem'.}
\item{dictionary}{The dictionary used for the model.}
\item{exog_data}{A tibble with the exogenous data used for the forecast.}
\item{exog_data_nowcast}{A tibble with the exogenous data used for the nowcasting.}
\item{nowcast_data}{A tibble with the nowcasted data.}
\item{args}{A list with the arguments used for the forecast.}
\item{full_forecast_data}{A tibble with the full forecast data, if available.}
}
}
\description{
Forecast OSEM model
}
\examples{
spec <- dplyr::tibble(
  type = c(
    "d",
    "d",
    "n"
  ),
  dependent = c(
    "StatDiscrep",
    "TOTS",
    "Import"
  ),
  independent = c(
    "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
    "GValueAdd + Import",
    "FinConsExpHH + GCapitalForm"
  )
)

\dontrun{
a <- run_model(specification = spec)
forecast_model(a)
}
}
