% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_model.R
\name{forecast_model}
\alias{forecast_model}
\title{Forecast Aggregate model}
\usage{
forecast_model(model, exog_predictions = NULL, n.ahead = 10, plot = TRUE)
}
\arguments{
\item{model}{An aggregate model of class 'aggmod'}

\item{exog_predictions}{A data.frame or tibble with values for the exogenous values. The number of rows of this data must be equal to n.ahead.}

\item{n.ahead}{Periods to forecast ahead}

\item{plot}{Logical. Should the result be plotted? Default is TRUE.}
}
\value{

}
\description{
Forecast Aggregate model
}
