% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_sensitivity.R
\name{forecast_sensitivity}
\alias{forecast_sensitivity}
\title{Evaluate the sensitivity of forecasts to varying exogenous values}
\usage{
forecast_sensitivity(model, size = 0.5, quiet = FALSE)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}

\item{size}{Numeric, must be larger than -1. The percentage value that the exogenous values will be multiplied with. Default is 0.5 (i.e. 50\%).}
}
\value{
A tibble with the original forecasts and the difference to the modified forecasts.
}
\description{
Evaluate the sensitivity of forecasts to varying exogenous values
}
