% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_download_variables.R
\name{load_locally}
\alias{load_locally}
\title{Load Locally}
\usage{
load_locally(to_obtain, inputdata_directory, quiet)
}
\arguments{
\item{to_obtain}{A data.frame as returned by \code{\link[=determine_variables]{determine_variables()}}.}

\item{inputdata_directory}{A path to .rds, .csv, or .xlsx input files in
which the data is stored. Can be \code{NULL}, in which case data is obtained
via download if possible.}

\item{quiet}{Logical with default = \code{FALSE}. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}
}
\value{
Returns a list with two named elements: \code{$df} stores the
downloaded data and \code{$to_obtain} the updated data.frame tracking which
variables still need to be obtained.
}
\description{
Loads data from a local directory
}
