% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_download_variables.R
\name{load_or_download_variables}
\alias{load_or_download_variables}
\title{Loads or downloads the necessary data for all modules}
\usage{
load_or_download_variables(
  specification,
  dictionary = NULL,
  primary_source = c("download", "local"),
  inputdata_directory = NULL,
  save_to_disk = NULL,
  quiet = FALSE,
  constrain.to.minimum.sample = TRUE
)
}
\arguments{
\item{specification}{A tibble or data.frame as returned by
\code{\link{check_config_table}}.}

\item{dictionary}{A tibble or data.frame storing the Eurostat variable code
in column 'eurostat_code'. If \code{download == TRUE} then the dictionary
also requires column named 'dataset_id' that stores the Eurostat dataset
id. When \code{NULL}, the \link[=dict]{default dictionary} is used.}

\item{primary_source}{A character of length 1 determining whether the
required variables are first obtained from \code{"download"} and then
\code{"local"}ly or vice versa.}

\item{inputdata_directory}{A path to .rds, .csv, or .xlsx input files in
which the data is stored. Can be \code{NULL}, in which case data is obtained
via download if possible.}

\item{save_to_disk}{A path to a directory where the final dataset will be
saved, including file name and ending. Not saved when \code{NULL}.}

\item{quiet}{Logical with default = \code{FALSE}. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}

\item{constrain.to.minimum.sample}{Logical. Should all data series be
constrained to the minimum data series? Default is \code{TRUE}.}
}
\description{
Either downloads the necessary data for all modules from Eurostat using
dataset ids or loads the files provided in the user-specified input directory
and selects the required variables.
}
