% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Visualize dependence between modules}
\usage{
network(model)
}
\arguments{
\item{model}{A model of class 'osem'}
}
\value{
Returns a network graph illustrating the dependence between the
different modules.
}
\description{
Visualize dependence between modules
}
