% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcasting}
\alias{nowcasting}
\title{Nowcast missing data for forecasting}
\usage{
nowcasting(model, exog_df_ready)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}
}
\value{
Returns a list with two full model objects. One contains the original model and one contains the nowcasted model.
}
\description{
Nowcast missing data for forecasting
}
