% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcasting}
\alias{nowcasting}
\title{Nowcast missing data for forecasting}
\usage{
nowcasting(model, exog_df_ready, frequency)
}
\arguments{
\item{model}{An aggregate model object of class 'aggmod'.}

\item{exog_df_ready}{The outcome of the \link[=forecast_exogenous_values]{forecast_exogenous_values} function, as prepared by the \link[=forecast_model]{forecast_model} function..}

\item{frequency}{Character string that indicates the frequency of the model. Must be compatible with the 'by' argument in \code{seq.Date()}.}
}
\value{
Returns a list with two full model objects. One contains the original model and one contains the nowcasted model.
}
\description{
Nowcast missing data for forecasting
}
