% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aggmod.forecast.R
\name{plot.aggmod.forecast}
\alias{plot.aggmod.forecast}
\title{Plot a Forecast Object of the Aggregate Model}
\usage{
\method{plot}{aggmod.forecast}(object)
}
\arguments{
\item{object}{An object of class aggmod.forecast, which is the output from the forecast_model function.}
}
\value{

}
\description{
Plot a Forecast Object of the Aggregate Model
}
