% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aggmod.forecast.R
\name{plot.aggmod.forecast}
\alias{plot.aggmod.forecast}
\title{Plot a Forecast Object of the Aggregate Model}
\usage{
\method{plot}{aggmod.forecast}(
  x,
  exclude.exogenous = TRUE,
  order.as.run = FALSE,
  interactive = FALSE,
  first_value = NULL,
  grepl_variables = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class aggmod.forecast, which is the output from the forecast_model function.}

\item{exclude.exogenous}{Logical. Should exogenous values be plotted? Default is FALSE.}

\item{order.as.run}{Logical. Should the plots be arranged in the way that the model was run? Default FALSE.}

\item{first_value}{Character. First date value to be shown. Must be a character value that can be turned into a date using as.Date() or NULL.}

\item{grepl_variables}{Regular Expression Character. Can be used to select variables to be plotted. Experimental feature so use with care.}

\item{...}{Further arguments (currently not in use).}
}
\description{
Plot a Forecast Object of the Aggregate Model
}
\examples{
spec <- dplyr::tibble(
type = c(
  "d",
  "d",
  "n"
),
dependent = c(
  "StatDiscrep",
  "TOTS",
  "Import"
),
independent = c(
  "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
  "GValueAdd + Import",
  "FinConsExpHH + GCapitalForm"
)
)
\donttest{
a <- run_model(specification = spec, dictionary = NULL,
inputdata_directory = NULL, primary_source = "download",
save_to_disk = NULL, present = FALSE)
plot(forecast_model(a))
}
}
