% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.osem.forecast.insample.R
\name{plot.osem.forecast.insample}
\alias{plot.osem.forecast.insample}
\title{Plot an Insample-Forecast of the OSEM Model}
\usage{
\method{plot}{osem.forecast.insample}(
  x,
  title = "OSEM Insample Hindcasts",
  first_date = NULL,
  first_date_insample_model = NULL,
  grepl_variables = NULL,
  grepl_method = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class osem.forecast.insample, which is the output from the \link{forecast_insample} function.}

\item{title}{Character. Title of the plot. Default is "OSEM Model Forecast".}

\item{first_date}{Character. First date value to be shown. Must be a character value that can be turned into a date using as.Date() or NULL.}

\item{first_date_insample_model}{A character string that will be converted to
a Date that indicates the first insample forecast to show in the plot. By default
(\code{NULL}) all insample forecasts will be shown.}

\item{grepl_variables}{Regular Expression Character. Can be used to select variables to be plotted. Experimental feature so use with care.}

\item{grepl_method}{A character string that will be used to filter the forecasting method.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot an Insample-Forecast of the OSEM Model
}
\examples{
spec <- dplyr::tibble(
type = c(
  "d",
  "d",
  "n"
),
dependent = c(
  "StatDiscrep",
  "TOTS",
  "Import"
),
independent = c(
  "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
  "GValueAdd + Import",
  "FinConsExpHH + GCapitalForm"
)
)
\donttest{
a <- run_model(specification = spec, dictionary = NULL,
primary_source = "download",
save_to_disk = NULL, present = FALSE)
plot(forecast_model(a))
}

}
