% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.osem.R
\name{plot.osem}
\alias{plot.osem}
\title{Plot OSEM Model}
\usage{
\method{plot}{osem}(
  x,
  title = "OSEM Model Output",
  exclude.exogenous = TRUE,
  order.as.run = FALSE,
  interactive = FALSE,
  first_date = NULL,
  grepl_variables = NULL,
  linewidth = 1,
  return.data = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class osem, which is the output from the \link{run_model} function.}

\item{title}{Character. Title of the plot. Default is "OSEM Model Output".}

\item{exclude.exogenous}{Logical. Should exogenous values be plotted? Default is FALSE.}

\item{order.as.run}{Logical. Should the plots be arranged in the way that the model was run? Default FALSE.}

\item{interactive}{Logical. Should the resulting plot be launched in an interactive way (the plotly package is required for this).}

\item{first_date}{Character. First date value to be shown. Must be a character value that can be turned into a date using as.Date() or NULL.}

\item{grepl_variables}{Regular Expression Character. Can be used to select variables to be plotted. Experimental feature so use with care.}

\item{linewidth}{Numeric. Linewidth argument for the plot. Default is 1.}

\item{return.data}{Logical. Do not return a plot but rather just the final dataset that has been created for the plot.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A ggplot object or a tibble/data.frame with the data used for plotting, depending on the value of \code{return.data}.
}
\description{
Plot the estimated OSEM Model
}
\examples{
spec <- dplyr::tibble(
type = c(
  "d",
  "d",
  "n"
),
dependent = c(
  "StatDiscrep",
  "TOTS",
  "Import"
),
independent = c(
  "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
  "GValueAdd + Import",
  "FinConsExpHH + GCapitalForm"
)
)
\donttest{
a <- run_model(specification = spec, dictionary = NULL,
primary_source = "download",
save_to_disk = NULL, present = FALSE)
plot(a)
}
}
