% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aggmod.R
\name{print.aggmod}
\alias{print.aggmod}
\title{Print output of an Aggregate Model}
\usage{
\method{print}{aggmod}(model)
}
\arguments{
\item{model}{An object of type 'aggmod'}
}
\description{
Print output of an Aggregate Model
}
