% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.osem.R
\name{print.osem}
\alias{print.osem}
\title{Print output of an OSEM Model}
\usage{
\method{print}{osem}(x, ...)
}
\arguments{
\item{x}{An object of type 'osem'}

\item{...}{Further arguments.}
}
\description{
Print output of an OSEM Model
}
