% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isat.R
\name{run_isat}
\alias{run_isat}
\title{Internal function to run and prepare isat estimation}
\usage{
run_isat(
  yvar,
  y.name,
  xvars,
  clean_data,
  ar = NULL,
  mc,
  saturation,
  saturation.tpval,
  max.block.size,
  pretest_steps,
  determine.blocksize
)
}
\arguments{
\item{yvar}{The y variable}

\item{y.name}{The name of the y variable.}

\item{xvars}{All x variables.}

\item{clean_data}{An input data.frame or tibble. Must be the output of
clean_data() to fit all requirements.}

\item{ar}{numeric vector or \code{NULL}. The AR order to be used.}

\item{mc}{Logical. Whether to use an intercept.}

\item{saturation}{Carry out Indicator Saturation using the 'isat' function
in the 'gets' package. Needs a character vector or string. Default is
'c("IIS","SIS")' to carry out Impulse Indicator Saturation and Step Indicator
Saturation. Other possible values are 'NULL' to disable or 'TIS' or Trend
Indicator Saturation. When disabled, estimation will be carried out using
the 'arx' function from the 'gets' package.}

\item{saturation.tpval}{The target p-value of the saturation methods (e.g.
SIS and IIS, see the 'isat' function in the 'gets' package). Default is 0.01.}

\item{max.block.size}{Integer. Maximum size of block of variables to be
selected over, default = 20.}

\item{pretest_steps}{Logical. Default is \code{FALSE}. This argument controls
whether isat should first be run for SIS in isolation before other saturation
methods are added (IIS, TIS). This can lead to better results if there are many
IIS identified at the end of the estimation sample/forecast origin. If \code{TRUE}
then \code{\link[gets]{isat}} is first carried out just for SIS (if activated using 'sis = TRUE'),
then the SIS breaks are pre-entered to another \code{\link[gets]{isat}} estimation but not selected over.
After both isat runs, a union model selection is done using  \code{\link[gets]{gets}}.}

\item{determine.blocksize}{Logical. Whether max.block.size or needs to be determined.}
}
\value{
An isat object.
}
\description{
Internal function to run and prepare isat estimation
}
