% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs the aggregate model}
\usage{
run_model(
  specification,
  dictionary = NULL,
  inputdata_directory = paste0(getwd(), "/data/raw"),
  filter_list = NULL,
  download = FALSE,
  save_to_disk = NULL,
  present = FALSE,
  quiet = FALSE,
  use_logs = c("both", "y", "x"),
  trend = TRUE,
  ardl_or_ecm = "ardl",
  max.ar = 4,
  max.dl = 4,
  saturation = c("IIS", "SIS"),
  saturation.tpval = 0.01,
  max.block.size = 20,
  gets_selection = TRUE,
  selection.tpval = 0.01
)
}
\arguments{
\item{specification}{A tibble or data.frame with three columns. Column names must be:
'type', 'dependent', and 'independent'. The column 'type' must contain for each row a
character of either 'd' (Identity) or 'n' (Definition - i.e. will be estimated).
The column 'dependent' must contain the LHS (Y variables) and the column named
'independent' containing the RHS (x variables separated by + and -).}

\item{dictionary}{A tibble or data.frame storing the Eurostat variable code
in column 'eurostat_code' and the model variable name in 'model_varname'.
If \code{download == TRUE} then the dictionary also requires a column named
'dataset_id' that stores the Eurostat dataset id. When \code{NULL}, the
\link[=dict]{default dictionary} is used.}

\item{inputdata_directory}{A path to .rds input files in which the data is
stored. Can be \code{NULL} if \code{download == TRUE}.}

\item{filter_list}{A named list with one entry per Eurostat code that
specifies the filter that is applied to choose the correct data series,
e.g. geographical restrictions or whether seasonally adjusted or not. See
\link{load_or_download_variables} for details.}

\item{download}{A logical value whether the data should be downloaded from
Eurostat or will be provided directly via \code{inputdata_directory}.}

\item{save_to_disk}{A path to a directory where the final dataset will be
saved, including the file name and ending. Not saved when \code{NULL}.}

\item{present}{A logical value whether the final aggregate model output
should be presented or not. NOTE: not implemented yet.}

\item{quiet}{Logical with default = FALSE. Should messages be displayed?
These messages are intended to give more information about the estimation
and data retrieval process.}

\item{use_logs}{To decide whether to log any variables. Must be one of "both", "y", or "x". Default is "both".}

\item{trend}{Logical. Should a trend be added? Default is TRUE.}

\item{ardl_or_ecm}{Either 'ardl' or 'ecm' to determine whether to estimate the model as an Autoregressive Distributed Lag Function (ardl) or as an Equilibrium Correction Model (ecm).}

\item{max.ar}{Integer. The maximum number of lags to use for the AR terms. as well as for the independent variables.}

\item{max.dl}{Integer. The maximum number of lags to use for the independent variables (the distributed lags).}

\item{saturation}{Carry out Indicator Saturation using the 'isat' function in the 'gets' package. Needes is a character vector or string. Default is 'c("IIS","SIS")' to carry out Impulse Indicator Saturation and Step Indicator Saturation. Other possible values are 'NULL' to disable or 'TIS' or Trend Indicator Saturation. When disabled, estimation will be carried out using the 'arx' function from the 'gets' package.}

\item{saturation.tpval}{The target p-value of the saturation methods (e.g. SIS and IIS, see the 'isat' function in the 'gets' package). Default is 0.01.}

\item{max.block.size}{Integer. Maximum size of block of variables to be selected over, default = 20.}

\item{gets_selection}{Logical. Whether general-to-specific selection using the 'getsm' function from the 'gets' package should be done on the final saturation model. Default is TRUE.}

\item{selection.tpval}{Numeric. The target p-value of the model selection methods (i.e. general-to-specific modelling, see the 'getsm' function in the 'gets' package). Default is 0.01.}
}
\value{
An object of class \link[=new_aggmod]{aggmod}, which is a named list
with four elements:
\describe{
\item{args}{A named list storing the user arguments for the aggregate
model.}
\item{module_order_eurostatvars}{The original specification with translated
variable names to Eurostat codes and arranged in order of estimation.}
\item{module_collection}{The above specification with two added columns
that store the model object for each module and the dataset used for
estimation, including fitted values for the dependent variable.}
\item{full_data}{A tibble or data.frame containing the complete original
data for the aggregate model and the fitted values of each module.}
}
}
\description{
Runs the aggregate model according to the given specification of modules.
}
\examples{
spec <- dplyr::tibble(
  type = c(
    "d",
    "d",
    "n"
  ),
  dependent = c(
    "StatDiscrep",
    "TOTS",
    "Import"
  ),
  independent = c(
    "TOTS - FinConsExpHH - FinConsExpGov - GCapitalForm - Export",
    "GValueAdd + Import",
    "FinConsExpHH + GCapitalForm"
  )
)

fa <- list(geo = "AT", s_adj = "SCA", unit = "CLV05_MEUR")
fb <- list(geo = "AT", s_adj = "SCA", unit = "CP_MEUR")
filter_list <- list("P7" = fa, "YA0" = fb, "P31_S14_S15" = fa,
"P5G" = fa, "B1G" = fa, "P3_S13" = fa, "P6" = fa)
\donttest{
a <- run_model(specification = spec, dictionary = NULL,
inputdata_directory = NULL, filter_list = filter_list,
download = TRUE, save_to_disk = NULL, present = FALSE)
}
}
