% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_module.R
\name{run_module}
\alias{run_module}
\title{Runs a module}
\usage{
run_module(module, data, classification, ...)
}
\arguments{
\item{module}{A row of the specification table.}

\item{data}{A tibble or data.frame containing the full data for the aggregate
model.}

\item{classification}{A data.frame with two columns: 'var' stores the
Eurostat variable code and 'class' the classification code (one of
\code{"n"}, \code{"d"}, or \code{"x"}) reflecting whether the variable is
endogenous by modelling, endogenous by identity/definition, or exogenous.
As returned by \link{classify_variables}.}
}
\value{
Returns a list with two named elements.
\describe{
\item{model}{For endogenously modeled modules (\code{type == "n"}), an
object of class \code{\link[gets:isat]{isat}} or
\code{\link[gets:arx]{arx}}. For identity/definition modules
(\code{type == "d"}), \code{NULL}.}
\item{data}{The data (including transformations) that was used to estimate
the model and the fitted values of the dependent variable.}
}
}
\description{
Runs a module in two different ways depending on whether the module is
endogenously modelled or endogenous by identity/definition.
}
