% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Creates and runs the Shiny App "shinyaggmodel"}
\usage{
run_shiny()
}
\description{
Opens a window for the user to interact with the app. The user can upload
an \link[=new_aggmod]{aggmod} object returned by \code{\link{run_model}} and
produce graphical and tabular output.
}
