% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Creates and runs the OSEM Shiny App}
\usage{
run_shiny(model = NULL)
}
\arguments{
\item{model}{The model object that is passed by the 'present_model()' function.}
}
\description{
Opens a window for the user to interact with the app. The user can upload
an \link[=new_osem]{osem} object returned by \code{\link{run_model}} and
produce graphical and tabular output.
}
